--- src/main/java/lab4/game/Position.java
@@ -1,10 +1,0 @@
-package lab4.game;
-
-/**
- * Represents a game board position
- * @param row
- * @param col
- */
-public record Position(Row row, Col col) {
-
-}
