+++ src/main/java/lab5/Main.java
@@ -1,7 +1,7 @@
-package tictactoe;
+package lab5.game;
 
-import tictactoe.game.TicTacToeGame;
-import tictactoe.ui.Console;
+import lab5.game.TicTacToeGame;
+import lab5.ui.Console;
 
 import static tictactoe.game.PlayerToken.O;
 import static tictactoe.game.PlayerToken.X;
