+++ src/main/java/lab5/players/Player.java
@@ -1,15 +1,14 @@
-package tictactoe.players;
+package lab5.players;
 
-import tictactoe.game.Board;
-import tictactoe.game.Position;
-import tictactoe.ui.Console;
+import lab5.game.Board;
+import lab5.game.Position;
+import lab5.ui.Console;
 
 /**
  * Represents a player in the game.
  *
- * TODO: Make this an abstract class with various subclasses for different types of players
  */
-public class Player {
+public abstract class Player {
 
     private String name;
 
@@ -23,18 +22,7 @@ public class Player {
      * Prompts the player to pick their next move.
      * @param currentBoard The current state of the game board
      * @return The position on the board where the player wants to place their token
-     *
-     * TODO: Make this an abstract method in an abstract class,
-     *          and use this implementation in a HumanPlayer subclass
-     */
-    public Position pickNextMove(Board currentBoard) {
-        while (true) {
-            var move = Console.promptForPosition(getName() + " pick your next move: ");
-            if (currentBoard.isEmptyAt(move)) {
-                return move;
-            } else {
-                Console.printAlert("That position is already taken.");
-            }
-        }
-    }
+     **/
+    public abstract void promptMove();
+
 }
